### 2024-09-16 #259 @editorialbot
[REVIEW]: Can you predict the future? A tutorial for the National Ecological Observatory Network Ecological Forecasting Challenge
**Submitting author:** @OlssonF (<a href="http://orcid.org/0000-0002-0483-4489">Freya Olsson</a>)
**Repository:** https://github.com/eco4cast/NEON-forecast-challenge-workshop
**Branch with paper.md** (empty if default branch): main
**Version:** v1.3.1
