### 2024-09-30 #259 @dlebauer
@acocac Could you please help me understand the version requirement "Does the release version given match the repository release?" - the latest release in the repository is v1.2.0, this issue lists it as v1.1.0. It isn't clear to me which version the proof is from, though I suspect any changes will at least bump the version.

--> I think it would be helpful to clarify this part of the checklist and reviewer instructions
