### 2024-10-04 #259 @skmorgane
@acocac I started my review by making sure I didn't have conflicts with the full author list but realized that two of the authors I have an association with that I should disclose: Carey and Thomas. We are coauthors on a large author line paper that was just accepted. I think this one falls under "co-authors but not really collaborating" in the guidelines and is probably not a concern. However, we are also collaborating on a proposal, again a large multi-investigator/multi-institutional effort, but we do interact a little more directly there. I think I can review this impartially, but if the association violates JOSE rules, I have two people I can recommend as reviewers who I know are not interacting with anyone on the author line. Just let me know how you would like to proceed
