### 2024-10-10 #259 @labarba
@skmorgane – Thank you for raising the potential conflict with some authors. Given that JOSE reviews are open, done in public, meant to be constructive, I am inclined to waive the COI in this instance. Our review process here is not a gatekeeping exercise, and authors recommend publication once all items in a review checklist are passed. Open peer review for the win!

Please go ahead and complete your review—and thank you kindly for your contributions 🙏
