### 2024-10-11 #259 @skmorgane
I have finished my review of the NEON-forecast-challenge-workshop by Olsson et al.

Overall I think this is a really useful contribution to the ecological forecasting teaching materials. Students in my class have often requested that we add something like this to our course but the energy it would take to design and implement has prevented us from doing that. I reviewed the materials with that in mind and found them to be well organized, easy to understand, and well focused on the essential skills students would need to be introduced to. I can easily see implementing these materials either as is or with minor modifications to reflect the focus of my class.

The only box I on the review sheet that I could not tick was the community engagement statement.

I found only minor issues with the materials and left issues in the source repo for the author's consideration.
