### 2024-10-14 #259 @acocac
> Thank you all for your time to review these materials! @acocac what are the steps that we should take for making revisions following the Reviewer comments? They having submitted pull requests and Github issues that we can address, but is there additional responses that should be completed? Thanks

Please address the reviewers comments, issues and PRs. Then, I'll double check if they're support the publication in JOSE. Thank you.
