### 2024-10-18 #259 @OlssonF
Thanks @acocac. We are also wanting to change the ownership of the repository to ensure long-term sustainability and use of the materials. We would like to move the materials from my personal organisation (@OlssonF) to the eco4cast organisation (https://github.com/eco4cast). Does it matter if this done before or after publication? I believe the transfer will not impact previous users interactions or forks etc.
Thanks
