### 2024-10-21 #259 @OlssonF
> @OlssonF please transfer to the organisation as soon as you can, and implement the changes suggested by the reviewers. Please notify here when everything is ready.

@acocac We have addressed the reviewers comments (from PR and Issues raised on the repository) and have completed the transfer to the new organisation. The materials are now hosted at [eco4cast/NEON-forecast-challenge-workshop](https://github.com/eco4cast/NEON-forecast-challenge-workshop). I believe everything is ready. Thanks!
