### 2024-10-22 #259 @acocac
@skmorgane, can you revisit your checklist and confirm if the author has satisfactorily addressed the community guidelines request? Would you recommend the work for publication?

Thank you!
