### 2024-10-22 #259 @OlssonF
> @OlssonF thanks for confirm the successful transfer of the repository. May I ask if you update the [paper](https://github.com/eco4cast/NEON-forecast-challenge-workshop/blob/main/paper/paper.md) too? It still points to the previous GitHub repository, https://github.com/OlssonF/NEON-forecast-challenge-workshop. Please commit/push the updated paper and release a version accordingly. Thank you 🙏

Thanks! We have updated the paper and release (https://github.com/eco4cast/NEON-forecast-challenge-workshop/releases/tag/v1.3.1)
