### 2024-10-31 #259 @OlssonF
The latest software release is v1.3.1 (https://github.com/eco4cast/NEON-forecast-challenge-workshop/releases/tag/v1.3.1) and the Zenodo DOI is https://doi.org/10.5281/zenodo.14018090. I have checked and updated the Zenodo archive title and authorship and updated the license to match the GitHub repository (MIT license)
Thanks @acocac
