### 2024-11-01 #259 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
