### 2024-12-26 #259 @labarba
Congratulations @OlssonF – your JOSE paper is published! 🚀

Huge thanks to our Editor: @acocac and Reviewers: @dlebauer, @skmorgane – your contributions make this possible 🙏
